removed thread pool

+++ include/cellworld_planner/tree_search.h
@@ -11,13 +11,13 @@ namespace cell_world::planner{
         const Cell &cell;
         const Graph &graph;
         void load();
-        Option &get_best_option(double exploration);
-        std::vector<double> rewards;
+        Option &get_best_option(float exploration);
+        std::vector<float> rewards;
         std::vector<unsigned int> counters;
         std::vector<Option> options;
-        std::vector<double> get_ucb1(double exploration);
+        std::vector<float> get_ucb1(float exploration);
         int best_option = Not_found;
-        void update_reward(double );
+        void update_reward(float);
     };
 
     struct History_step : json_cpp::Json_object {
